+++ src/style/create.css
@@ -97,3 +97,9 @@ em{
 label{
     font-size: 20px ;
 }
+
+button.disabled:hover{
+    text-shadow: none;
+    color:rgba(16,16,1 , 0.3) ;
+    cursor:not-allowed
+}
